Fix webpack warning about iconv-loader
ems-client has a dependency on node-fetch which is only used for Node.js. Webpack throw a warning when it sees the iconv-loader.js file in the encoding dependency of node-fetch. We only use the browser target of ems-client so node-fetch is not used here. So we can suppress this warning by telling webpack to ignore the iconv-loader.js file.