Remove ci jobs for older branches (#376)
These jobs compile and publish the static assets for older releases of landing page based on their respective branches. We do not expect to make any changes to these releases so there is no reason to keep the CI jobs. If we do need to update an older release, we can manually compile and publish the assets.